app-store-api/requirements.txt
FastAPI
uvicorn
sqlalchemy
asyncpg
pydantic
passlib[bcrypt]
python-jose
alembic
databases
email-validator
pydantic-settings
psycopg2
motor
pymongo-migrate